+++ .circleci/config.yml
@@ -1,0 +1,57 @@
+version: 2
+
+defaults: &defaults
+  working_directory: ~/workspace
+  docker:
+    - image: node:12.10.0-alpine
+
+only-publish-tags: &only-publish-tags
+  filters:
+    tags:
+      only: /^v?\d+\.\d+\.\d+$/
+    branches:
+      ignore: /.*/
+
+jobs:
+  build:
+    <<: *defaults
+    steps:
+      - run: apk add ca-certificates git openssh
+      - checkout
+      - restore_cache:
+          keys:
+            - node-modules-{{ checksum "package.json" }}
+            - node-modules-
+      - run: npm install
+      - save_cache:
+          paths:
+            - node_modules
+          key: node-modules-{{ checksum "package.json" }}
+      - run: npx roboter
+      - persist_to_workspace:
+          root: ~/workspace
+          paths:
+            - .
+  publish:
+    <<: *defaults
+    steps:
+      - run: apk add ca-certificates
+      - attach_workspace:
+          at: ~/workspace
+      - run: echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > ~/workspace/.npmrc
+      - run: npx roboter build
+      - run: npm publish
+
+workflows:
+  version: 2
+  untagged-build:
+    jobs:
+      - build
+  tagged-build:
+    jobs:
+      - build: *only-publish-tags
+      - publish:
+          <<: *only-publish-tags
+          requires:
+            - build
+          context: Default
