+++ .dependabot/config.yml
@@ -1,0 +1,18 @@
+version: 1
+
+update_configs:
+  - package_manager: "javascript"
+    directory: "/"
+    update_schedule: "live"
+    default_reviewers:
+    - "thenativeweb/wolkenkit-dev"
+    default_assignees:
+    - "goloroden"
+    - "yeldiRium"
+    default_labels:
+    - "Dependencies"
+    allowed_updates:
+    - match:
+        dependency_type: "direct"
+        update_type: "all"
+    version_requirement_updates: "increase_versions"
