+++ .npmignore
@@ -1,0 +1,3 @@
+.circleci
+.dependabot
+test
