+++ package.json
@@ -10,13 +10,18 @@
     {
       "name": "Matthias Wagler",
       "email": "matthias.wagler@thenativeweb.io"
+    },
+    {
+      "name": "Hannes Leutloff",
+      "email": "hannes.leutloff@thenativeweb.io"
     }
   ],
-  "main": "dist/measureTime.js",
+  "main": "build/lib/measureTime.js",
+  "types": "build/lib/measureTime.d.ts",
   "dependencies": {},
   "devDependencies": {
-    "assertthat": "2.0.1",
-    "roboter": "3.0.4"
+    "assertthat": "4.0.2",
+    "roboter": "8.1.3"
   },
   "repository": {
     "type": "git",
